move function to api
